+++ ddprof-lib/src/main/cpp/stackWalker.inline.h
@@ -47,12 +47,4 @@ inline void fillFrame(ASGCT_CallFrame& frame, FrameTypeId type, int bci, jmethod
     frame.method_id = method;
 }
 
-inline jmethodID getMethodId(VMMethod* method) {
-    if (!StackWalkValidation::inDeadZone(method) && StackWalkValidation::aligned((uintptr_t)method)
-            && SafeAccess::isReadableRange(method, VMMethod::type_size())) {
-        return method->validatedId();
-    }
-    return NULL;
-}
-
 #endif // _STACKWALKER_INLINE_H
